requirements.txt
flask
flask_cors
uvicorn
pydantic
python-dotenv
google-genai
gunicorn==23.0.0 
tavily-python